Unified request classes
