--- src/main/java/com/progress/tracking/wrapper/trello/pojo/CreateCardAttachmentRequest.java
@@ -1,55 +1,0 @@
-package com.progress.tracking.wrapper.trello.pojo;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-public class CreateCardAttachmentRequest {
-
-    @SerializedName("name")
-    @Expose
-    private String name;
-
-    @SerializedName("url")
-    @Expose
-    private String url;
-
-    @SerializedName("token")
-    @Expose
-    private String token;
-    
-    @SerializedName("key")
-    @Expose
-    private String key;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getUrl() {
-        return url;
-    }
-
-    public void setUrl(String url) {
-        this.url = url;
-    }
-
-    public String getToken() {
-        return token;
-    }
-
-    public void setToken(String token) {
-        this.token = token;
-    }
-
-    public String getKey() {
-        return key;
-    }
-
-    public void setKey(String key) {
-        this.key = key;
-    }
-}
